added a build schema step now that we have deleted the file from source code and also fixed issue 32
