+++ server/src/diskdb.ts
@@ -1,11 +1,9 @@
 import { promises as fs } from 'fs';
 import * as uuid from 'uuid';
-import { Repository } from './types';
+import { Repository, isValidProject, isValidQuickChecklist } from './types';
 import {
   ProjectModel,
   QuickChecklistModel,
-  isValidProject,
-  isValidQuickChecklist,
 } from '../../frontend/src/types';
 import * as lockfile from 'proper-lockfile';
