+++ server/src/dynamodb.ts
@@ -1,11 +1,9 @@
 import * as AWS from 'aws-sdk';
 import * as uuid from 'uuid';
-import { Repository } from './types';
+import { Repository, isValidProject, isValidQuickChecklist } from './types';
 import {
   ProjectModel,
   QuickChecklistModel,
-  isValidProject,
-  isValidQuickChecklist,
 } from '../../frontend/src/types';
 import { ServiceConfigurationOptions } from 'aws-sdk/lib/service';
